Fix for the cert cache ets tables not being cleaned correctly
Two fixes are included in this patch, first off the destination of a mesage has been
fixed to be delivered to the ssl_manager process, instead of the calling processes.
The second fix is more of a patch for odd behaviour when the messages are delivered.
Occationally the PEM cache would be reset, and the Ref associated with a MD5 hash
would change to reflect that the cache was rest. This would cause the case statement
on line 326 to fail as no clauses to match, because the Ref had chagned. The catch on
line 323 is for processes requesting a cache be cleaned for data that does not exist any more.